[ddk][serial] New serial protocol for DDK
This new serial protocol comes in two parts:

ZX_PROTOCOL_SERIAL is a remotable protocol intended for use by clients.
It provides a socket for the actual data transfer.

ZX_PROTOCOL_SERIAL_DRIVER is a low level in-process protocol intended to be implemented
by UART drivers on SOC platforms.
It provides synchronous read and write calls, as well as support for waiting
for signals on a handle for readable/writeable state changes.

In addition, this adds support for serial ports to be distributed as resources
to platform devices, similar to how we support GPIOs and I2C channels in the platform bus.
The platform bus driver will use the ZX_PROTOCOL_SERIAL_DRIVER protocol to interface with
SOCS UART drivers and will expose these via the ZX_PROTOCOL_SERIAL protocol
to platform device drivers.

It is expected that the ZX_PROTOCOL_SERIAL will also be implemented by other serial drivers,
like intel-serialio.

Change-Id: Ife83a350276c2f68801ab66057727b01076cb0f1